requirements.txt
Flask==3.0.3
gunicorn==22.0.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
Pillow==10.4.0